requirements.txt
open3d==0.18.0
diffusers==0.27.2
einops==0.7.0
fire==0.6.0
h5py==3.10.0
imageio==2.34.0
kornia==0.7.2
pytorch-lightning==2.1.3
lpips==0.1.4
matplotlib==3.8.4
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
pillow==10.2.0
pytorch-msssim==1.0.0
pyyaml==6.0.1
rembg==2.0.56
scikit-image==0.21.0
scikit-learn==1.4.0
scipy==1.13.0
tensorboardx==2.6.2.2
timm==0.9.12
torchmetrics==1.4.0
torchtyping==0.1.4
tqdm==4.66.2
transformers==4.25.1
trimesh==4.3.2
wandb==0.16.2
xformers==0.0.23.post1
tensorboard==2.16.2
open_clip_torch==2.24.0
streamlit==1.35.0
invisible-watermark==0.2.0
pypng
#third_party/diff-surfel-rasterization
#git+https://github.com/openai/CLIP.git